Enhance quantization, serialization, and gradient handling
- Updated quantization methods in AWQ and GPTQ to use half-to-even rounding with std::nearbyint for consistency with PTQ and QAT.
- Improved fake quantization gradient handling by widening per-element contributions to Float32 before reduction to prevent overflow.
- Added endianness handling in SafeTensors serialization to ensure compatibility across different architectures, including byte-swapping for big-endian hosts.
- Enhanced tensor serialization and deserialization to support complex types and ensure correct endianness.
- Modified training loop to report NaN for validation loss when no validation loader is provided, preventing false improvements in early stopping.
- Implemented gradient scaling in gradient accumulation to maintain effective learning rates across micro-batches.
- Adjusted spectral normalization to handle complex weights correctly during power iteration.
- Fixed output padding handling in ConvTranspose1d to ensure correct tensor dimensions when output_padding > 0.

+++ src/backends/cuda/kernels/conv2d.cu
@@ -1984,7 +1984,10 @@ auto conv_transpose2d_forward_kernel(
         const float* bias_ptr = has_bias ? bias->data<float>() : nullptr;
         float* output_ptr = output.data<float>();
 
-        conv_transpose2d_forward_kernel_impl<float, double><<<grid, block, 0, stream>>>(
+        // Accumulate F32 in float to match the CPU reference and the ROCm/Vulkan
+        // conv_transpose kernels. A double accumulator made CUDA the lone outlier,
+        // drifting O(sqrt(N)) ULPs from the other backends past tolerance (F038).
+        conv_transpose2d_forward_kernel_impl<float, float><<<grid, block, 0, stream>>>(
             input_ptr, weight_ptr, bias_ptr, output_ptr,
             batch, in_channels, in_h, in_w,
             out_channels, out_h, out_w,
